Refactor and remove duplication
